Student Full Detail App For Beginner
This is another full-detail project for beginners. Here, you will provide the input and get the stored data in ArrayOfObject. Input may be the student's name or roll number. 

I have stored all data in Array using the object. Means object in array. By this method you will be able to store large data in objects using an array.