valobj.js (get values of an object non-case sensitive)
You can get the values of an object using a prompt or input field. Either key is uppercase or lowercase. No, issue.  non-case sensitive.

Non-Case sensitive means, if you put a key (student name or etc) in the prompt box for getting the value of the targeted key. There is no issue. You're putting the key in lowercase or uppercase. You will get the proper result.