keyvalobj.js (get a property of an object using key or value non-case sensitive)
Through this code, you will be able to find a property of an object (data of a student) using key or value non-case sensitive.

Non-case sensitive means, you can get any property using a key or value. It does not matter if you are putting the query in the prompt is lowercase or uppercase or capitalize.

You will get the same result.